Use more ideomatic Mojo and unify integer width with C.
Mojo prefers to make use of the SIMD type for small collections. In this case, it allocates a i32x16 block because it has to be a power of 2, but only the first 10 slots are used as with the original algorithm. The data type has been switched to Int32 instead of Int, because in Mojo Int actually maps to ssize_t, which is int64_t on x86-64 Linux. 

The usage of @parameter is also idemomatic for loops with compile-time-known iteration counts.